add VideoPlayer component parent && rename VideoParentInner as child
